🎨  create middleware for logging
- added TODO to add unique request id
- added TODO to move logging middleware to ignition
- see https://github.com/TryGhost/Ghost/issues/7116
